contrib/fetch/requirements.txt
requests[security]
#pandas
python-dateutil
python-dotenv
psycopg[binary]
